Small changes in architecture: Logic of methods moved from repository to service, Added check if user id. @Evgen

+++ src/main/java/com/example/userstorage/service/UserServiceImpl.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class UserServiceImpl implements UserService{
+public class UserServiceImpl implements UserService {
 
     @Autowired
     private UserRepository repository;
@@ -18,13 +18,13 @@ public class UserServiceImpl implements UserService{
     @Override
     @SneakyThrows
     public User create(UserRequest userRequest) {
-        if(userRequest ==null){
+        if (userRequest == null) {
             throw new Exception("Wrong data");
         }
-        int idUser = repository.getAll().size()+1;
+        int idUser = repository.getAll().size() + 1;
 
-        User user = new User(idUser,userRequest.getFirstName(), userRequest.getLastName(), userRequest.getAge());
-        if(repository.getByLastName(userRequest.getLastName()) != null){
+        User user = new User(idUser, userRequest.getFirstName(), userRequest.getLastName(), userRequest.getAge());
+        if (getByLastname(userRequest.getLastName()) != null) {
             throw new Exception("User with such lastname already exists");
         }
         return repository.create(user);
@@ -33,10 +33,13 @@ public class UserServiceImpl implements UserService{
     @Override
     @SneakyThrows
     public User update(User user) {
-        if(user ==null){
+        if (user == null) {
             throw new Exception("Wrong data");
         }
-        return repository.update(user);
+        if (repository.getAll().contains(user)) {
+            return repository.update(user);
+        }
+        throw new Exception("No such user");
     }
 
     @Override
@@ -45,18 +48,33 @@ public class UserServiceImpl implements UserService{
     }
 
     @Override
+    @SneakyThrows
     public User getById(Integer id) {
-        return repository.getById(id);
+        if (id <= repository.getAll().size() && repository.getAll().get(id - 1) != null) {
+            return repository.getById(id);
+        }
+        throw new Exception("No such user");
     }
 
     @Override
     public User getByLastname(String lastname) {
-        return repository.getByLastName(lastname);
+        if (lastname != null) {
+            for (User u : repository.getAll()) {
+                if (u.getLastName().equals(lastname)) {
+                    return u;
+                }
+            }
+        }
+        return null;
     }
 
     @Override
+    @SneakyThrows
     public void delete(Integer id) {
-        repository.delete(id);
-
+        if (id <= repository.getAll().size() && repository.getAll().get(id-1) != null) {
+            repository.delete(id);
+            return;
+        }
+        throw new Exception("No such user");
     }
 }
